fix: rename xml to comicxml
This should make this xml package "unique" and prevent require clashing
with other packages that have a similar name.

Fixes: https://github.com/NightQuest/comicmeta.koplugin/issues/8
Relates-To: https://github.com/koreader/koreader/issues/14277
