+++ lib/comicxml.lua
